requirements.txt
autopep8==2.3.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
flake8==7.1.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-logging==3.11.0
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.2
grpcio-status==1.62.2
idna==3.7
importlib_metadata==8.0.0
mccabe==0.7.0
opentelemetry-api==1.26.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pyflakes==3.2.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
wrapt==1.16.0
zipp==3.19.2
